Added reverse prop option + changed uart3 to uart2 on magnum board

+++ sw/inc/utils.h
@@ -27,7 +27,7 @@ void wait_ms(uint32_t t);
 float uint32_to_float(uint32_t x);
 uint32_t int32_to_uint32(int32_t x);
 int32_t uint32_to_int32(uint32_t x);
-void dshot_encode(volatile uint16_t* val, volatile uint32_t buf[16], _Bool telemetry);
+void dshot_encode(volatile uint16_t* val, volatile uint8_t buf[16], _Bool telemetry);
 float expo(float lin);
 float arcsin(float sin_val);
 float sinus(float angle);
